Use language-specific type names in Java and CS static-final/constant definitions #3480
In `src/ide/frames/fields/constant-value-field.ts` change `getElanType` to return
the language-specific name for the type (eg "double" for Java) and not the Elan name (eg "Float").
In the interests of minimum code change, I have not changed the function name.
Update all the expected output in 24 Java and CS test files, Int -> int etc.
This makes the rendered code more like valid Java or CS.

+++ test/files/pathfinder.java
@@ -307,13 +307,13 @@ enum Algorithm {dijkstra, aStar, heuristic}
 
 // 'infinity' need only be larger than longest possible route
 
-static final Float infinity = 2000.0; // constant
+static final double infinity = 2000.0; // constant
 
 static final String algPrompt = "Enter 'a' for A-star, 'd' for Dijkstra, 'h' for Heuristic"; // constant
 
-static final Int lightBlue = 0x80abff; // constant
+static final int lightBlue = 0x80abff; // constant
 
-static final Int orange = 0xe97132; // constant
+static final int orange = 0xe97132; // constant
 
 static Algorithm getAlgFromLetter(String letter) { // function
   var alg = Algorithm.heuristic;
